fix(chat): rename error message component
