+++ apps/chat/src/components/Common/ApplicationWizard/QuickAppView.tsx
@@ -25,7 +25,7 @@ import { TemperatureSlider } from '@/src/components/Chat/Temperature';
 import { ApplicationWizardFooter } from '@/src/components/Common/ApplicationWizard/ApplicationWizardFooter';
 import { DropdownSelector } from '@/src/components/Common/DropdownSelector';
 import { Field } from '@/src/components/Common/Forms/Field';
-import { withErrorMessage } from '@/src/components/Common/Forms/FieldError';
+import { withErrorMessage } from '@/src/components/Common/Forms/FieldErrorMessage';
 import { FieldTextArea } from '@/src/components/Common/Forms/FieldTextArea';
 import { withLabel } from '@/src/components/Common/Forms/Label';
 import { CustomLogoSelect } from '@/src/components/Settings/CustomLogoSelect';
